Improve highlighting
