+++ wikichange/src/content/markContent.js
@@ -1,5 +1,5 @@
 import { HighlightType, HIGHLIGHT_TYPE } from "./enums";
-import { escapeRegex } from "./oustideCode";
+import { cleanText, escapeRegex } from "./cleanText";
 
 /**
  * 
@@ -20,6 +20,8 @@ const markContentHelper = (_text, _mark, _remove_mark) => {
      * and the last is the end index of the item found
      */
     const textMatching = (context) => {
+        cleanText(context);
+        console.log(context.highlight);
         const {content_before, highlight, content_after } = context;
         const start = [..._text.matchAll(new RegExp(escapeRegex(highlight), 'g'))];
         if (start.length > 0){
