--- wikichange/src/content/oustideCode.js
@@ -1,11 +1,0 @@
-/**
- * https://stackoverflow.com/questions/3561493/is-there-a-regexp-escape-function-in-javascript/3561711#3561711
- * Code from stack overflow to escape all special character from string to be used in a regex
- * @param {string} string to be used in regex
- * @returns {string} an escaped string that will not cause regex to read any special characters
- */
-function escapeRegex(string) {
-    return string.replace(/[/\-\\^$*+?.()|[\]{}]/g, '\\$&');
-}
-
-export { escapeRegex };
